fix: prevent panic when parsing a URL with no path (#55)
* chore: sate clippy

* fix: prevent panic when parsing a URL with no path

Prior to this change `parse` would panic at `src/lib.rs:203:29` with
`index out of bounds: the len is 0 but the index is 0` if given an input
like "git:".

* feat: improve error handling

This removes a seemingly redundant and not very illuminating error
variant. It also ensures that the error message from the `Url` crate is
displayed when displaying `UrlParseError`.